Exercise 4.20 completed, invalid delete operation if the user is not the creator

+++ part4/controllers/blog.js
@@ -3,14 +3,6 @@ const jwt = require('jsonwebtoken');
 const Blog = require('../models/blog');
 const User = require('../models/user');
 
-const getTokenFrom = (request) => {
-  const authorization = request.get('authorization');
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7);
-  }
-  return null;
-};
-
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { blogs: 0 });
   response.json(blogs);
@@ -22,8 +14,7 @@ blogRouter.post('/', async (request, response, next) => {
     if (!request.body.title && !request.body.url) { return response.status(400).end(); }
     const { body } = request;
 
-    const token = getTokenFrom(request);
-    const decodedToken = jwt.verify(token, process.env.SECRET);
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
     if (!decodedToken.id) {
       return response.status(401).json({ error: 'token missing or invalid' });
     }
@@ -39,21 +30,33 @@ blogRouter.post('/', async (request, response, next) => {
         user: user._id,
       },
     );
-  
+
     const savedBlog = await blog.save();
     user.blogs = user.blogs.concat(savedBlog._id);
     await user.save();
     return response.status(201).json(savedBlog);
   } catch (error) {
-    return next(error)
+    return next(error);
   }
 });
 /* eslint-enable */
 
 blogRouter.delete('/:id', async (request, response) => {
-  const { id } = request.params;
-
   try {
+    const { id } = request.params;
+    const { token } = request;
+
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' });
+    }
+    const user = await User.findById(decodedToken.id);
+    // console.log(user);
+
+    if (!(user.blogs.find((blog) => blog._id.toString() === id.toString()))) { return response.status(400).json({ error: 'Not the creatror' }); }
+    // console.log(decodedToken);
+
     await Blog.findByIdAndDelete(id);
     response.status(204).end();
   } catch (e) {
